Initial codebase for PSModule.Sodium (#10)
Sodium binary module with an IsolatedAssemblyLoadContext for Sodium.Core
(libsodium).

## Description

An implementation of [Loading through .NET Core Assembly Load
Contexts](https://github.com/PSModule/Sodium/compare/main...vercellone:Sodium:feature/IsolatedAssemblyLoadContext?expand=1)
to provide a ConvertTo-SodiumEncryptedString CmdLet as a foundation for
GitHub secret encryption.

- Inspired by [PSSodium](https://github.com/TylerLeonhardt/PSSodium) and
[PowerShell-Yayaml](https://github.com/jborean93/PowerShell-Yayaml).
- Based largely on
[ModuleDependencyIsolationExample](https://github.com/rjmholt/ModuleDependencyIsolationExample).
- Unlike PSSodium, this implementation targets pwsh 7.4+ / net8.0 (LTS)
with no windows powershell.exe compatibility.
- src/PSModule.Sodium/build.ps1 is excluding linux-musl-* runtime
assemblies, but that could easily change.
- src/PSModule.Sodium/build.ps1 publishes to ./lib/net8.0. The output
path could easily be changed.
- ./lib/** has been pushed to the repo. It could be excluded from the
repo and built just-in-time by GitHub workflows and/or vscode task(s).
But, IMHO including it simplifies local development scenarios.

I suspect you'll want to PSModule-ize this a bit. But look it over and
let me know if this is as useful a foundation as I think it is for
GitHub secret encryption.

## Type of change

- [x] 🚀 [Feature]

## Checklist

- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas

---------

Co-authored-by: Vercellone, Jason <Jason.Vercellone@mkcorp.com>