🌟 [Major]: Align function naming with `libsodium` (#22)
## Description

This pull request attempts to align naming of functions, parameters and
outputs to be closer with `libsodiium` terminology. This includes
significant changes to the encryption and decryption functions,
transitioning from using the term "SodiumEncryptedString" to
"SodiumSealedBox".

- Fixes #20 

### Changes to function names and parameters

* `ConvertFrom-SodiumEncryptedString` renamed to
`ConvertFrom-SodiumSealedBoxString`, with updates to function name,
parameters, and documentation to use "SealedBox" instead of
"EncryptedString". Also updates the parameter from `Secret` to
`SealedBox` to better align with `libsodium`.
* `ConvertTo-SodiumEncryptedString` renamed to
`ConvertTo-SodiumSealedBoxString`, with updates to function name,
parameters, and documentation to use "SealedBox" instead of
"EncryptedString". Also uipdates the parameter from `Secret` to
`Message` to better align with `libsodium`.

### Updates to examples

*
[`examples/CreateGitHubSecret.ps1`](diffhunk://#diff-b7943e4c6e438d9b3c4c861bdc8fe422a1d5e5df32be6f6aab3e34edb699a149L4-R4):
Updated the encryption function call to use `ConvertTo-SodiumSealedBox`
instead of `ConvertTo-SodiumEncryptedString`.
*
[`examples/GitHubSecretService.ps1`](diffhunk://#diff-fe6c0fb0a8d439a5e6ed90bbaa6394857934aaf8c8771ef862546e69ca9437a6L25-R25):
Updated encryption and decryption function calls to use
`ConvertTo-SodiumSealedBox` and `ConvertFrom-SodiumSealedBox`
respectively.
[[1]](diffhunk://#diff-fe6c0fb0a8d439a5e6ed90bbaa6394857934aaf8c8771ef862546e69ca9437a6L25-R25)
[[2]](diffhunk://#diff-fe6c0fb0a8d439a5e6ed90bbaa6394857934aaf8c8771ef862546e69ca9437a6L46-R50)

### Documentation updates

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L40-R67):
Updated examples to use "Sealed Boxes encryption" and changed function
calls to `ConvertTo-SodiumSealedBox` and `ConvertFrom-SodiumSealedBox`.

### Test updates:
*
[`tests/Module.Tests.ps1`](diffhunk://#diff-a94f0e90abaaff044e3cb0327ce3a63a10cd61551269a414e4dc40ad2af8a951L2-R75):
Added new tests for encryption and decryption using
`ConvertTo-SodiumSealedBox` and `ConvertFrom-SodiumSealedBox`, including
error cases for invalid keys.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [x] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas